Concatenate the main class after its dependencies.
grunt-uglify tries to be smart about the order it concatenates scripts in, and puts function declerations before statements, to try and make sure that everything is defined before anything attempts to call it.

The typical pattern for defining Backbone objects is a statement, though, so we need to manually push the main class to the end of the generated file.