Generate a unique ID for links to avoid exact duplicates.
Only exact duplicates are removed; links with identical URLs but different titles were intentionally left in the result set because we don't know which title is the best, and even if we did, the user may be searching for a specific title and would be confused if they couldn't find it.

Murmur was chosen as the hashing algorithm because of its speed and uniqueness.

See https://programmers.stackexchange.com/questions/49550/which-hashing-algorithm-is-best-for-uniqueness-and-speed
See http://stackoverflow.com/a/22429679/450127