Escape post titles with esc_html() instead of Underscore.js.
`get_the_title()` applies `the_title`, which calls `wptexturize()` on the title. `wptexturize()` converts certain characters to their HTML entities, so we need to be able to display the title as HTML, rather than text. Otherwise, titles like `Flannery O'Conner` would be displayed as `Flannery O&#8217;Conner`.

To make the title safe to display as HTML, it's escaped with `esc_html()` before it's passed to the JavaScript side.