Ignore the Grunt-compiled files for developer convenience.
They don't need to be tracked in the development repository, since they'll be automatically generated by Grunt. Tracking them just makes development less convenient because any changes are duplicated in `git diff`, and have to be `git add`'ed, etc.

Not tracking them means that the plugin won't work if someone just `clone`'s it and doesn't realize they need to run `grunt watch`, so a `README.md` was added.