Fetch content index even if it can't be cached.
Blink/WebKit-based browsers don't expose `window.caches` over HTTP, even though they should. Previously those users would get an error that the index can't be fetched, and a request to update to a newer browser. We can still fetch the index even if we can't cache it, though; that's a performance hit, but a much better UX overall.